# Point In Polygon
https://www.codewars.com/kata/530265044b7e23379d00076a/train/javascript

The goal of this kata is to implement a function called `pointInPoly`.
`pointInPoly` is called with a polygon represented as a series of points
as the first argument, and a single point at the second argument. Each
point is represented as a set of `[x, y]` coordinates, where both `x` and
`y` are numbers. `pointInPoly` should return `true` if the point is
within the defined polygon, and `false` if it is not.

The kata description points out several assumptions we can make about
the inputs:
1) The polygon will be a valid simple polygon. That is, it will have at
least three points, none of its edges will cross each other, and exactly
two edges will meet at each vertex.
2) In the tests, the point will never fall exactly on an edge of the
polygon.

And lastly, although the description never explicitly says so, I'm
assuming that the points in the polygon are given in order; each point
shares an edge with the next.

This initial commit sets up the project to build out this solution.
`index.js` is where we'll build our solution, and `test/index.js` is
where we'll do our testing.