# The Simplest Test
The first thing we do when we're solving a problem like this is to write
a test. Because we're implementing a function given to us, we already
know what our interface will look like (`pointInPoly`), so we can
immediately write a test for it.

Our first test asserts that a point at the origin (`[0, 0]`) is within a
simple triangle with points at `[-1, -1]`, `[1, -1]`, and `[0, 1]`.

After writing this test, our test suite complains that `pointInPoly is
not defined`. This is quickly remediated by importing `pointInPoly` into
our test file and then exporting it from `index.js`.

After exporting the empty `pointInPoly` function, the test suite shouts
that it `expected undefined to be true`. A quick fix to bring us back to
green is to simply have our new `pointInPoly` function return `true`.