# Fleshing Out a Strategy
We knew that our initial `pointInPoly` solution was naive and
incomplete. Returning `true` for all cases obviously wasn't going to
work. But what do we do instead? How do we even begin to tackle this
problem?

Thankfully, I know from my days of programming video games that a
[simple
test](https://en.wikipedia.org/wiki/Point_in_polygon#Ray_casting_algorithm)
for checking if a point lies within a polygon is to send a ray outward
from that point. If it intersects with the lines of the polygon an odd
number of times, the point lies within the polygon. Otherwise, it's
outside.

Since we're in a green state, we can do a little refactoring and
implement a high level version of this solution. We basically want to
count the number of intersections between our imaginary ray and our
polygon:

    let intersections = countIntersections(poly, point);

And then return `true` if that count is odd, and `false` if it's even:

    return !!(intersections %2);

After making these changes, our test suite is complaining that
`countIntersections` does not exist, so let's quickly define it and have
it return `1` to bring us back to green.