# Rethinking Interfaces
After some thinking and soul-searching, I decided I wasn't happy with
where I was going with our previous solution. `countIntersections` was
really an internal method. Outside of the context of our
point-in-polygon problem, a function called `countIntersections` that
takes in a `poly` and a `point` really just doesn't make any sense.

Because it was an internal method, I was hesitant to write tests for it.
What if I eventually want to refactor and fundamentally change or remove
`countIntersections` alltogether? Additionally, `countIntersections`
would most likely call other methods would be even more contextually
dependant and awkward to test.

I needed a cleaner solution.

After reconsidering the problem, I noticed that we're dealing with a few
really solid abstractions. The most apparent is a `polygon`. If we
implement a generic `polygon`, we'd be able to cleanly specify what we
want from our `pointInPoly` method:

    function pointInPoly(poly, point) {
        return polygon(poly).surrounds(point);
    }

Additionally, by breaking `polygon` into its own abstraction, we can
freely build and test its interface to our heart's content.

With this in mind, I wrote a new set of tests that describe `polygon`.
The first tests looks nearly identical to our `pointInPoly` tests and
checks if a polygon surrounds a point.

A dummy implementation of `polygon.surrounds` simply returns `true`.