# Another Abstraction
Now we turn out attention to the `intersections` function. Honestly,
I haven't been looking forward to this. This still seems like a big
piece of functionality and I'd like to break it down into simpler
pieces, if possible.

If we're not afraid of implementing another abstraction, a simple
implementation of `intersections` could be written like this:

    return lines().filter(line => line.intersects(ray)).length;

In plain english, this reads as "return the number of lines in this
polygon that intersect the given ray".

I like this solution. To make it a reality, I created a new set of tests
for our new `line` abstraction, and a dummy implementation of `line` and
`line.intersects`.