# Finishing Translate
Adding a more complicated test of `translate` showed that we needed a
better solution. Thankfully, it wasn't a huge step to come up with the
final form of the function.

The `translate` function returns a new `polygon` where every point in
the polygon has been incremented by the provided `x` and `y` values.