# Finishing Line
Adding another test to `line` shows that we need a better solution.

Determining if a line intersects with a ray is a fairly well-covered
problem. I used [this blog
post](https://rootllama.wordpress.com/2014/06/20/ray-line-segment-intersection-test-in-2d/)
as a guide for implementing my solution. Be sure to check it out if
you're not sure what's going on here.