# Integrating Lines
Finally, all we need to do to complete our solution is to build our
`polygon.lines` function. This function should transform the set of
`_points` that were used to define a polygon into a set of `line`
objects.

After implementing a test for `polygon.lines` we're able to come up with
a solution. Don't forget that the last point in the
polygon must connect back to the first!