# Restating Point-in-Polygon
Now that we've defined our `polygon`, we can restate our high-level
point-in-polygon test. We want to `translate` our polygon so that the
point we've been given acts like a ray from the origin. Next we want to
count the number of intersections that ray makes with the newly
translated polygon:

    let intersections = translate([-x, -y]).intersections([0, 1]);

Lastly we want to return `true` from `surrounds` if `intersections` is
odd:

    return !!(intersections % 2);

After making these changes, our test suite complains about `translate`
and `intersections` not being defined.

The fastest way to get us back to green is to have `translate` return a
new `polygon`, and have `intersections` return `1`.